+++ src/main/java/br/com/decolatech/projetoDecolatech/domain/models/TaskList.java
@@ -10,7 +10,7 @@ public class TaskList {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
     private String title;
     private Date CreationDate;
     private boolean arquived;
@@ -21,7 +21,7 @@ public class TaskList {
     public TaskList() {
     }
 
-    public TaskList(int id, String title, Date creationDate, boolean arquived, List<Task> tasks) {
+    public TaskList(Long id, String title, Date creationDate, boolean arquived, List<Task> tasks) {
         this.id = id;
         this.title = title;
         CreationDate = creationDate;
@@ -29,11 +29,11 @@ public class TaskList {
         this.tasks = tasks;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
